% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{calculus_index}
\alias{calculus_index}
\title{Calculate calculus index}
\source{
{Greene, T.R., Kuba, C.L., Irish, J.D. 2005. Quantifying calculus: A suggested new approach for recording an important indicator of diet and dental health.} \doi{10.1016/j.jchb.2005.02.002}
}
\usage{
calculus_index(.data, ..., .score = score, .method = c("simple", "greene2005"))
}
\arguments{
\item{.data}{Data frame in long format, with one tooth surface per row, and one calculus score per surface.}

\item{...}{can be used to add arguments to \code{\link[dplyr]{group_by}}. By default .add = T.}

\item{.score}{numeric. Calculus score for each surface. Defaults to 'score' as provided by \link{dental_longer}.}

\item{.method}{string. Using \code{.method = "greene2005"} will apply \code{dental_filter()}
and add \code{quadrant} to grouping variables (according to Greene et al. (2005)).
\code{.method = "simple"} (default) just takes user input and no filter.}
}
\value{
The default returns a data frame with the supplied groups and calculations
for number of surfaces scored, number of teeth scored, sum of scores, and the
calculus index.
}
\description{
Convenience function to calculate dental calculus index directly from a
\code{dental_longer()} output.
}
\details{
Calculate the calculus index for each quadrant of an individual based
on the method outlined in Greene et al. (2005). The original method calculates
a score for each quadrant of the dentition, so it is recommended to group by
quadrant, either in the function or using \code{\link[dplyr]{group_by}} above in a pipe (see examples).
}
\examples{
library(dplyr)
mb11_calculus_long <- mb11_calculus \%>\%
  dental_longer(-id, names_sep = "_") \%>\%
  dental_join()
# example adding groups before call to function
mb11_calculus_long \%>\%
  group_by(id, quadrant) \%>\%
  calculus_index()

# example adding groups in call to function
mb11_calculus_long \%>\%
  calculus_index(id, quadrant)

# example of the Greene et al. 2005 method
mb11_calculus_long \%>\%
  calculus_index(.method = "greene2005")
}
