% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caries.R
\name{caries_ratio}
\alias{caries_ratio}
\title{Calculate caries ratio}
\usage{
caries_ratio(
  .data,
  ...,
  .caries = score,
  .no_lesion = NULL,
  .lesion_sep = NULL,
  .method = c("location", "standards", "count")
)
}
\arguments{
\item{.data}{Data frame in long format, with one tooth per row, and a column with the caries count for each tooth.}

\item{...}{can be used to add arguments to \code{\link[dplyr]{group_by}}. By default .add = T.}

\item{.caries}{Caries count variable. Defaults to 'count'.}

\item{.no_lesion}{string used to indicate how the absence of lesions were scored.}

\item{.lesion_sep}{string used to indicate how multiple lesions on a tooth/surface are separated (if applicable).}

\item{.method}{string. The method used to score caries lesions. Supported methods are 'location', 'standards', and 'count'.
See 'Details' for more information.}
}
\value{
Returns a data frame with caries ratio per group.
}
\description{
Convenience function to calculate dental caries ratio directly from a
\code{dental_longer()} output.
}
\details{
Calculate the caries ratio for each tooth type. Calculated as the number
of caries divided by the number of visible surfaces.

Multiple methods are available for recording carious lesions. The function supports
using a string to indicate the location of a lesion(s) on each tooth/surface
('location'),
the method outlined in Standards (Buikstra and Ubelaker 1994) ('standards'),
and an integer indicating the number of lesions on a tooth/surface ('count').
The method is #' passed to \code{caries_reduce()}, which reduces the recorded score
to presence/absence. The presence/absence is then summed to get a count for the
number of lesions on each tooth/surface.
}
\examples{
library(dplyr)
mb11_caries_long <- mb11_caries \%>\%
  dental_longer(-id, rm_char = "t")
# example adding groups before call to function
mb11_caries_long \%>\%
  group_by(class) \%>\%
  caries_ratio(.no_lesion = "none", .lesion_sep = ";")

# example adding groups in call to function
mb11_caries_long \%>\%
  caries_ratio(class, .no_lesion = "none", .lesion_sep = ";")
}
