% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_notation}
\alias{convert_notation}
\title{Convert between tooth notations}
\usage{
convert_notation(data, tooth_cols, from, to)
}
\arguments{
\item{data}{A data frame to convert}

\item{tooth_cols}{the column(s) to convert. If the data are in long format,
i.e., one row per tooth, this should be a single column. If the data are in
wide format, this should be a (\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}) selection of columns.}

\item{from}{string. the tooth notation from the original data.}

\item{to}{string. The tooth notation you wish to convert to.}
}
\description{
Function to switch between tooth notations.
}
