% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_height_recons_2d.R
\name{crown_height_recons_2d}
\alias{crown_height_recons_2d}
\title{Reconstructing slightly worn human teeth}
\usage{
crown_height_recons_2d(
  image_url,
  tooth_type = c("UI", "UC", "UP", "UM", "LI", "LC", "LP", "LM"),
  color_points = "red",
  color_regression_line = "blue",
  save_svg = c("yes", "no"),
  file_name_svg = "tooth_recons.svg"
)
}
\arguments{
\item{image_url}{Set the URL of the image of the tooth you would like to reconstruct}

\item{tooth_type}{set the tooth type}

\item{color_points}{select the color of the points of the regression equation ggplot. Default = "red"}

\item{save_svg}{select if the ggplot should be saved or not as SVG format: yes / no}

\item{file_name_svg}{if save_svg = "yes", this argument indicates the name of the resulting svg file. Default = "tooth_recons.svg"}

\item{interval}{set the interval of the regression: prediction / confidence}
}
\value{
a ggplot object with the regression equation placed appropriately in the region of interest
}
\description{
Reconstructing slightly worn human teeth
}
\details{
Reconstructing slightly worn human teeth
}
\examples{
crown_height_recons_2d("images/premolar.png",
tooth_type = "LP",
save_svg = "yes",
color_points = "red",
color_regression_line = "blue",
file_name_svg = "tooth_recons.svg")
}
