% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{dental-indices}
\alias{dental-indices}
\alias{dental_index}
\alias{dental_ratio}
\title{Calculate a dental index or ratio}
\usage{
dental_index(data, score = score)

dental_ratio(data, count = count)
}
\arguments{
\item{data}{long format data frame with tooth position information and either lesion counts or scores.}

\item{score}{Scores of a dental lesion (e.g. caries counts or dental calculus score).}

\item{count}{Counts of a dental lesion (e.g. caries counts or dental calculus score).}
}
\value{
Returns a data frame containing the grouping variable(s), the number of
teeth per group, the lesion count or sum of scores, and the calculated index
or ratio.
}
\description{
Calculate a dental index, such as a calculus index, or ratio, such as caries
ratio.
}
\details{
Function to calculate the index or ratio of dental diseases.
\code{dental_ratio()} is just a wrapper for \code{dental_index()}
where the output column 'index' is changed to 'count', and 'score_sum' to 'ratio'.
}
\section{Functions}{
\itemize{
\item \code{dental_index()}: calculate dental disease index

\item \code{dental_ratio()}: calculate dental disease ratio

}}
\examples{
library(dplyr)
library(tidyr)
# dental calculus index
mb11_calculus \%>\%
  dental_longer(-id, rm_char = "t", names_sep = "_") \%>\%
  group_by(class) \%>\% # calculate index per tooth class
  dental_index()

}
