% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dental_arcade_mapping}
\alias{dental_arcade_mapping}
\title{Dental arcade mapping data}
\format{
\subsection{\code{dental_arcade_mapping}}{

A data frame with 32 rows and 2 columns:
\describe{
\item{tooth}{string. Tooth with text notation.}
\item{geometry}{sfc. Simple feature geometry list-column for mapping teeth to geom_sf().}
...
}
}
}
\source{
https://github.com/ZandraFagernas/dental_arcade/blob/v0.1/dental_arcade_empty_mapping.jpg
}
\usage{
dental_arcade_mapping
}
\description{
Used to map each tooth to geom_sf() to create a dental plot.
}
\keyword{datasets}
