% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{dental_filter}
\alias{dental_filter}
\title{Dentition filter}
\usage{
dental_filter(data, score = score)
}
\arguments{
\item{data}{long format data frame with tooth position information and either lesion counts or scores.}

\item{score}{Scores of a dental lesion (e.g. caries counts or dental calculus score).}
}
\value{
Returns a filtered data frame and a warning if any rows were removed.
}
\description{
Checks if the data frame satisfies the criterion that minimally, each surface
must be scorable on at least one side of the dentition to obtain the index,
then filters out groups that don't satisfy.
}
