% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dental_join}
\alias{dental_join}
\title{Produce detailed tooth position information from tooth notation}
\usage{
dental_join(data, notation = NULL, rm_char = NULL)
}
\arguments{
\item{data}{A data frame containing one column with a unique identifier and
multiple columns with calculus scores from each surface of a tooth.}

\item{notation}{string. Which notation convention is used to name the columns
with dental scores. Options are "FDI", "standards1994", "text". See the
\code{tooth_notation} object for more details.}

\item{rm_char}{string. Optional argument to remove any characters from tooth
name columns, e.g., if the FDI number was prefixed with a letter (since it's not
great practice to have column names in a spreadsheet starting with a number).}
}
\value{
Returns the original data frame along with detailed information about
the tooth including, region (maxilla, mandible),
position (anterior, posterior), side, class (incisor, canine, etc.), type,
(first incisor, second molar, etc.), and quadrant.
}
\description{
Produce detailed tooth position information from tooth notation
}
