% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dental_plyr.R
\name{dental_recode}
\alias{dental_recode}
\title{Recode tooth notation}
\usage{
dental_recode(data, col, from, to)
}
\arguments{
\item{data}{A data frame to convert}

\item{col}{the column to convert. Must contain a tooth notation (e.g., FDI, standards).}

\item{from}{string. the tooth notation of the original data.}

\item{to}{string. The tooth notation you wish to convert to.}
}
\description{
Function to switch between tooth notations.
}
