% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb11_calculus}
\alias{mb11_calculus}
\title{Middenbeemster dental calculus data}
\format{
\subsection{\code{mb11_calculus}}{

A data frame with 41 rows and 97 columns:
\describe{
\item{id}{Skeletal ID}
\item{t11:t48_buc/lin/ipx}{Calculus score per tooth surface. Variable naming:
t<tooth number (FDI notation)>_\if{html}{\out{<surface>}}.
buc = buccal; lin = lingual, ipx = interproximal}
...
}
}
}
\source{
https://doi.org/10.5281/zenodo.7649151
}
\usage{
mb11_calculus
}
\description{
Dental calculus scores collected from Middenbeemster (MB11), a rural 19th century
archaeological site in the Netherlands.
}
\keyword{datasets}
