% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb11_caries}
\alias{mb11_caries}
\title{Middenbeemster dental caries data}
\format{
\subsection{\code{mb11_caries}}{

A data frame with 41 rows and 33 columns:
\describe{
\item{id}{Skeletal ID}
\item{t11:t48}{Caries lesion location(s) per tooth. Multiple scores separated
by ';'. Variable naming: t<tooth number (FDI notation)>.
mes = mesial, dis = distal, buc = buccal; lin = lingual, occ = occlusal,
root = root, crown = large lesion of indeterminate origin}
...
}
}
}
\source{
https://doi.org/10.5281/zenodo.7649151
}
\usage{
mb11_caries
}
\description{
Dental caries lesions collected from Middenbeemster (MB11), a rural 19th century
archaeological site in the Netherlands.
}
\keyword{datasets}
