% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb11_demography}
\alias{mb11_demography}
\title{Middenbeemster sample demographics}
\format{
\subsection{\code{mb11_demography}}{

A data frame with 41 rows and 97 columns:
\describe{
\item{id}{Skeletal ID}
\item{age}{Age category of the individual. eya = early young adult (18-24);
lya = late young adult (25-34); ma = middle adult (35-49); old = old adult (50+).}
\item{sex}{Biological sex of individual estimated using osteological methods.
f = female; pf = probable female; pm = probable male; m = male.}
...
}
}
}
\source{
https://doi.org/10.5281/zenodo.7649151
}
\usage{
mb11_demography
}
\description{
Age and sex of the individuals with dental calculus scored in \code{mb11_calculus}.
}
\keyword{datasets}
