% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perikymata_prediction.R
\name{perikymata_prediction}
\alias{perikymata_prediction}
\title{Predict perikymata counts in worn human teeth}
\usage{
perikymata_prediction(
  data,
  tooth = c("I", "C", "P", "M"),
  n_deciles = c("1", "2", "3"),
  decimals = 1
)
}
\arguments{
\item{data}{a data.frame with the perikymata counts in the deciles that are preserved in the tooth.}

\item{tooth}{a character representing the tooth type to predict perikymata. Only four tooth types are available: "I" Incisors, "C" Canines, "P" Premolars, "M" Molars}

\item{n_deciles}{a character representing the number of deciles to predict perikymata. Only three options are available: "1" predicts perikymata in DC1, "2" predicts perikymata in DC1 and DC2, "3" predicts perikymata in DC1, DC2 and DC3.}

\item{decimals}{a numeric to indicate the number of decimals for the perikymata number. By default it is set to 1.}
}
\value{
a data.frame with the perikymata prediction per decil.
}
\description{
Predict perikymata counts in worn human teeth
}
\details{
Reconstruct crown heights and estimate perikymata counts to calculate crown formation times. These variables are key for evolutionary and paleobiological studies.
}
\examples{
# Creation of a dummy dataframe of a premolar with perikymata from DC2 to DC10
perikymata_example <- data.frame(
  DC2  = c(10),
  DC3  = c(10),
  DC4  = c(11),
  DC5  = c(11),
  DC6  = c(14),
  DC7  = c(17),
  DC8  = c(19),
  DC9  = c(21),
  DC10 = c(21))

# Running the function to predict perikymata number in the first decile (DC1) of this dummy premolar.
perikymata_prediction(perikymata_example,
                      tooth = "P",
                      n_deciles = "1",
                      decimals = 1)
}
