% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tooth_notation}
\alias{tooth_notation}
\title{Common naming schemes in dental anthropology}
\format{
\subsection{\code{tooth_notation}}{

A data frame with 32 rows and 9 columns:
\describe{
\item{FDI}{Fédération dentaire internationale naming convention}
\item{standards}{Standards naming (Buikstra and Ubelaker 1994)}
\item{text}{Text naming: [UL][LR][ICPM][1-3] = upper/lower; left/right; incisor/canine/premolar/molar; tooth type (1-4). Example: Upper left lateral incisor = ULI2}
\item{region}{if the tooth is from the maxilla or mandible}
\item{position}{if the tooth is anterior or posterior}
\item{side}{if the tooth is from the left or right side}
\item{class}{if the tooth is incisor, canine, premolar, or molar}
\item{type}{if the tooth is central/lateral incisor, canine, first/second premolar, or first/second/third molar}
\item{quadrant}{Which quadrant of the mouth the tooth is from. Upper/lower anterior/posterior.}
...
}
}
}
\source{
Buikstra, J. E., & Ubelaker, D. H. (1994). Standards for data collection from human skeletal remains. Arkansas Archaeological Survey Research Series No. 44. Fayetteville: Arkansas Archaeological Survey.
}
\usage{
tooth_notation
}
\description{
Used to convert between tooth naming conventions that are common in dental
anthropology.
}
\keyword{datasets}
