% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dental_longer}
\alias{dental_longer}
\title{Convert a data frame containing dental scores to long format and add tooth
position variables.}
\usage{
dental_longer(
  data,
  cols,
  ...,
  notation = c("FDI", "standards1994", "text"),
  rm_char = NULL
)
}
\arguments{
\item{data}{A data frame containing one column with a unique identifier and
multiple columns with calculus scores from each surface of a tooth.}

\item{cols}{Which columns to pivot into longer format. These should be the columns
containing the scores for each tooth/surface.}

\item{...}{additional arguments passed on to \code{\link[tidyr]{pivot_longer}}.}

\item{notation}{string. Which notation convention is used to name the columns
with dental scores. Options are "FDI", "standards1994", "text". See the
\code{tooth_notation} object for more details.}

\item{rm_char}{string. Optional argument to remove any characters from tooth
name columns, e.g., if the FDI number was prefixed with a letter (since it's not
great practice to have column names in a spreadsheet starting with a number).}
}
\value{
Returns a longer data frame with one row per tooth or tooth surface,
with columns for tooth number, region (maxilla, mandible),
position (anterior, posterior), side, class (incisor, canine, etc.), type, and
quadrant.
}
\description{
This function is a wrapper for \code{pivot_longer()} that converts a wide format
data frame to a long format with tooth names in a \code{tooth} column
(and \code{surface}, if applicable) and the score in a \code{score} column. It also
attaches additional columns with information about the type of tooth and the
position of the tooth in the mouth (type, class, region, etc.).
}
\details{
If there are multiple columns per tooth, i.e., if teeth were scored per surface
and variables are named something like \if{html}{\out{<tooth number>}}_\if{html}{\out{<surface name>}},
\code{names_sep = "_"} (or whatever character was used to separate tooth number from
tooth surface) should be passed to pivot_longer via \code{...}. If no such
separator was used, use instead \verb{names_pattern = <regex>}. See
\code{\link[tidyr]{pivot_longer}} for more details.
}
\examples{
# pivot calculus scores per surface (all columns except id)
  # 't' prefix needs to be removed from column names
mb11_calculus \%>\%
  dental_longer(-id, rm_char = "t", names_sep = "_")

# pivot caries scores per tooth
mb11_caries \%>\%
  dental_longer(-id, rm_char = "t")
}
