% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perikymata_prediction.R
\name{perikymata_prediction}
\alias{perikymata_prediction}
\title{Predict perikymata counts in worn human teeth}
\usage{
perikymata_prediction(
  data,
  tooth = c("I", "C", "P", "M"),
  n_decils = c("1", "2", "3"),
  decimals = 1
)
}
\arguments{
\item{data}{a data.frame with the perikymata counts in the decils that are preserved in the tooth.}

\item{tooth}{a character representing the tooth type to predict perikymata. Only four tooth types are available: "I" Incisors, "C" Canines, "P" Premolars, "M" Molars}

\item{n_decils}{a character representing the number of decils to predict perikymata. Only three options are available: "1" predicts perikymata in DC1, "2" predicts perikymata in DC1 and DC2, "3" predicts perikymata in DC1, DC2 and DC3.}

\item{decimals}{a numeric to indicate the number of decimals for the perikymata number. By default it is set to 1.}
}
\value{
a data.frame with the perikymata prediction per decil.
}
\description{
Predict perikymata counts in worn human teeth
}
\details{
Reconstruct crown heights and estimate perikymata counts to calculate crown formation times. These variables are key for evolutionary and paleobiological studies.
}
